Remove Flux version of reader/follow-button
